requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-extensions==2.1.6
djangorestframework==3.9.2
idna==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
